🔨 refactor: api 경로 및 id 타입 수정

+++ src/main/java/com/x1/frans/user/query/dto/SearchFranchiseUserDTO.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @Setter
 public class SearchFranchiseUserDTO extends SearchUserDTO {
 
-    private Integer franchiseId;
+    private Long franchiseId;
     private String franchiseCode;
     private String franchiseName;
     private String franchiseAddress;
